Extract opaque logic to service
