+++ src/Api/Auth/Controllers/OpaqueKeyExchangeController.cs
@@ -1,7 +1,7 @@
 ﻿using Bit.Api.Auth.Models.Request.Opaque;
 using Bit.Api.Auth.Models.Response.Opaque;
+using Bit.Core.Auth.Services;
 using Bit.Core.Services;
-using Bitwarden.OPAQUE;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -11,54 +11,33 @@ namespace Bit.Api.Auth.Controllers;
 [Authorize("Web")]
 public class OpaqueKeyExchangeController : Controller
 {
-    private readonly IUserService _userService;
-    private readonly BitwardenOpaqueServer _bitwardenOpaque;
+    private readonly IOpaqueKeyExchangeService _opaqueKeyExchangeService;
+    IUserService _userService;
 
     public OpaqueKeyExchangeController(
+        IOpaqueKeyExchangeService opaqueKeyExchangeService,
         IUserService userService
     )
     {
+        _opaqueKeyExchangeService = opaqueKeyExchangeService;
         _userService = userService;
-        _bitwardenOpaque = new BitwardenOpaqueServer();
     }
 
     [HttpPost("~/opaque/start-registration")]
     public async Task<OpaqueRegistrationStartResponse> StartRegistration([FromBody] OpaqueRegistrationStartRequest request)
     {
         var user = await _userService.GetUserByPrincipalAsync(User);
-        var registrationRequest = _bitwardenOpaque.StartRegistration(request.CipherConfiguration, null, System.Convert.FromBase64String(request.RegistrationRequest), user.Id.ToString());
-        var message = registrationRequest.registrationResponse;
-        var serverSetup = registrationRequest.serverSetup;
-        // persist server setup
-        var sessionId = Guid.NewGuid();
-        SessionStore.RegisterSessions.Add(sessionId, new RegisterSession() { SessionId = sessionId, ServerSetup = serverSetup, cipherConfiguration = request.CipherConfiguration });
-        return new OpaqueRegistrationStartResponse(sessionId, System.Convert.ToBase64String(message));
+        var result = await _opaqueKeyExchangeService.StartRegistration(System.Convert.FromBase64String(request.RegistrationRequest), user, request.CipherConfiguration);
+        return new OpaqueRegistrationStartResponse(result.Item1, System.Convert.ToBase64String(result.Item2));
     }
 
 
     [HttpPost("~/opaque/finish-registration")]
     public async Task<String> FinishRegistration([FromBody] OpaqueRegistrationFinishRequest request)
     {
-        await Task.Run(() =>
-        {
-            var registerSession = SessionStore.RegisterSessions[request.SessionId];
-            var registrationFinish = _bitwardenOpaque.FinishRegistration(registerSession.cipherConfiguration, System.Convert.FromBase64String(request.RegistrationUpload));
-            Console.WriteLine("Registration Finish: " + registrationFinish);
-        });
+        await Task.Run(() => { });
         return "";
     }
 
 }
 
-public class RegisterSession
-{
-    public Guid SessionId { get; set; }
-    public byte[] ServerSetup { get; set; }
-    public CipherConfiguration cipherConfiguration { get; set; }
-}
-
-public class SessionStore()
-{
-    public static Dictionary<Guid, RegisterSession> RegisterSessions = new Dictionary<Guid, RegisterSession>();
-    public static Dictionary<Guid, byte[]> LoginSessions = new Dictionary<Guid, byte[]>();
-}
